[#6][front] convert shadow layout style to scss
