+++ front/app/src/components/layouts/default-layout/default-layout-sidebar.ce.scss
@@ -1,0 +1,7 @@
+default-layout-sidebar + default-layout-main {
+  padding: 3rem 0 0;
+
+  @media (min-width: 992px) {
+    padding: 0 0 0 15rem;
+  }
+}
