+++ front/app/src/views/settings.js
@@ -1,5 +1,5 @@
-import '@/components/layouts/default-layout-sidebar.ce.js';
-import '@/components/layouts/default-layout-main.ce.js';
+import '@/components/layouts/default-layout/default-layout-sidebar.ce.js';
+import '@/components/layouts/default-layout/default-layout-main.ce.js';
 
 const template = `
   <default-layout-sidebar></default-layout-sidebar>
