+++ front/app/src/views/view-login.ce.js
@@ -1,10 +1,10 @@
-import '@/components/layouts/auth-layout.ce.js';
+import '@/components/layouts/auth-layout/auth-layout.ce.js';
 import { redirectTo } from '@/router.js';
 
 class ViewSignIn extends HTMLElement {
   connectedCallback() {
     this.innerHTML = `
-      <login-layout>
+      <auth-layout>
         <h1 class="fw-bold py-2 mb-4">
           <span class="text-bicolor">Log In</span>
         </h1>
@@ -39,7 +39,7 @@ class ViewSignIn extends HTMLElement {
             </div>
           </div>
         </form>
-      </login-layout>
+      </auth-layout>
     `;
 
     // this.querySelector('a[data-link="/forget-pass"]').addEventListener('click', (event) => {
